Move React Query setup into main app boilerplate
