+++ templates/boilerplate/jest.setup.js
@@ -2,6 +2,8 @@ import '@testing-library/jest-native/extend-expect';
 import { configure } from '@testing-library/react-native';
 import mockSafeAreaContext from 'react-native-safe-area-context/jest/mock';
 import mockBackHandler from 'react-native/Libraries/Utilities/__mocks__/BackHandler.js';
+import server from 'src/test/server';
+import queryClient from 'src/util/api/queryClient';
 
 beforeEach(() => {
   jest.clearAllMocks();
@@ -34,6 +36,18 @@ jest.mock('@react-native-async-storage/async-storage', () =>
 );
 
 jest.mock('react-native-keyboard-aware-scroll-view');
+
+// listen with MSW server. Individual tests can pass mocks to 'render' function
+beforeAll(() => server.listen({ onUnhandledRequest: 'error' }));
+afterAll(() => server.close());
+
+beforeEach(() => {
+  server.resetHandlers();
+});
+
+afterEach(() => {
+  queryClient.clear();
+});
 
 // configure debug output for RN Testing Library
 // is way too verbose by default. Only include common
