+++ templates/boilerplate/package.json
@@ -23,10 +23,13 @@
     "@react-navigation/bottom-tabs": "^6.5.20",
     "@react-navigation/native": "^6.1.10",
     "@react-navigation/native-stack": "^6.9.18",
+    "@tanstack/react-query": "^5.32.1",
+    "axios": "^1.6.8",
     "expo": "^50.0.17",
     "expo-status-bar": "~1.11.1",
     "jest": "^29.3.1",
     "jest-expo": "~50.0.2",
+    "msw": "^2.2.14",
     "react": "18.2.0",
     "react-native": "0.73.6",
     "react-native-keyboard-aware-scrollview": "^2.1.0",
@@ -40,6 +43,7 @@
     "@thoughtbot/eslint-config": "^1.0.2",
     "@types/jest": "^29.5.12",
     "@types/react": "~18.2.73",
+    "@types/react-test-renderer": "^18.0.7",
     "babel-jest": "^29.7.0",
     "create-belt-app": "^0.4.0",
     "eslint": "^8.56.0",
