+++ templates/boilerplate/src/__tests__/App.test.tsx
@@ -4,6 +4,13 @@ import mock from 'src/test/mock';
 import { renderApplication } from 'src/test/render';
 
 test('renders', async () => {
+  // We would not normally recommend fake timers, but the tests are currently
+  // throwing a "not wrapped in act" warning after this test finishes. One
+  // option is to put a `await waitForUpdates()` at the end of the test, but
+  // fake timers also work here until we find a better solution. The stack trace
+  // seems to point to React Navigation bottom tabs.
+  jest.useFakeTimers();
+
   const mocks = [mockCoffees()];
 
   renderApplication({ mocks });
