+++ templates/boilerplate/src/test/sleep.ts
@@ -1,0 +1,5 @@
+export default function sleep(ms: number) {
+  return new Promise<void>((resolve) => {
+    setTimeout(resolve, ms);
+  });
+}
