+++ templates/boilerplate/src/util/api/queryClient.ts
